Bug: 뱃지뷰따로생성해서 버그 해결,뱃지 Date생성,위젯이름변경
