+++ PJ2T12_Again12/PJ2T12_Again12/Views/HomeView.swift
@@ -36,7 +36,6 @@ struct HomeView: View {
                                 Text("하고싶으면")
                                     .font(.Hel17Bold)
                                     .foregroundStyle(.defaultBlack)
-                                    .bold()
                                 Spacer()
                             }
                             VStack {
