+++ PJ2T12_Again12/PJ2T12_Again12/Views/StatusView.swift
@@ -12,31 +12,22 @@ import Charts
 struct StatusView: View {
     @FetchRequest(sortDescriptors: [SortDescriptor(\.date)], predicate: NSPredicate(format: "isTodo == true")) var todoList: FetchedResults<Todo>
     @FetchRequest(sortDescriptors: [SortDescriptor(\.date)], predicate: NSPredicate(format: "isTodo == false")) var wantTodoList: FetchedResults<Todo>
-    @State private var todoBadges = ["FirstF", "Todo10F", "Wantto10F"]
-    @State private var friendBadges = [ "FriendF", "AlertF", "FightingF" ]
+    
     @State private var selectedbadge = ""
-    @AppStorage("firstWantTodoIt") var firstWantTodoIt: Int = 0
+
     
     @State var Data = [
-        //        DataType(doType: "todo", data: [TodoOverView(date: 1, done: 1, undone: 1)]),
-        //        DataType(doType: "wantTodo", data: [TodoOverView(date: 1, done: 1, undone: 1)])
+
         
         DataType(doType: "해야하면", data: TodoOverView.todoMonth),
         DataType(doType: "하고싶으면", data: TodoOverView.wantTodoMonth)
     ]
-    
-    @State var showMedals = false
-    @State var settingsDetent = PresentationDetent.medium
+
     
     init() {
         UINavigationBar.appearance().largeTitleTextAttributes = [.foregroundColor: UIColor.brown]
     }
-    
-    let columns = [
-        GridItem(.flexible()),
-        GridItem(.flexible()),
-        GridItem(.flexible())
-    ]
+
     
     var body: some View {
         NavigationView {
@@ -86,71 +77,7 @@ struct StatusView: View {
                         Spacer()
                     }
                     .padding(.horizontal)
-                    LazyVGrid(columns: columns, spacing: 20) {
-                        Button {
-                            selectedbadge = firstWantTodoIt > 0 ? "FirstT" : "FirstF"
-                            showMedals = true
-                        } label: {
-                            Image(firstWantTodoIt > 0 ? "FirstT" : "FirstF")
-                                .resizable()
-                                .scaledToFit()
-                                .frame(width: 75)
-                                .padding(12)
-                                .overlay {
-                                    RoundedRectangle(cornerRadius: 20)
-                                        .stroke((Color.BorderBrown).opacity(0.32), lineWidth: 2)
-                                }
-                        }
-                        Button {
-                            selectedbadge = firstWantTodoIt > 9 ? "Todo10T" : "Todo10F"
-                            showMedals = true
-                        } label: {
-                            Image(firstWantTodoIt > 9 ? "Todo10T" : "Todo10F")
-                                .resizable()
-                                .scaledToFit()
-                                .frame(width: 75)
-                                .padding(12)
-                                .overlay {
-                                    RoundedRectangle(cornerRadius: 20)
-                                        .stroke((Color.BorderBrown).opacity(0.32), lineWidth: 2)
-                                }
-                        }
-                        Button {
-                            selectedbadge = firstWantTodoIt > 9 ? "Wantto10T" : "Wantto10F"
-                            showMedals = true
-                        } label: {
-                            Image(firstWantTodoIt > 9 ? "Wantto10T" : "Wantto10F")
-                                .resizable()
-                                .scaledToFit()
-                                .frame(width: 75)
-                                .padding(12)
-                                .overlay {
-                                    RoundedRectangle(cornerRadius: 20)
-                                        .stroke((Color.BorderBrown).opacity(0.32), lineWidth: 2)
-                                }
-                        }
-                        ForEach(friendBadges, id: \.self) { badge in
-                            Button {
-                                selectedbadge = badge
-                                showMedals = true
-                            } label : {
-                                Image(badge)
-                                    .resizable()
-                                    .scaledToFit()
-                                    .frame(width: 75)
-                                    .padding(12)
-                                    .overlay {
-                                        RoundedRectangle(cornerRadius: 20)
-                                            .stroke((Color.BorderBrown).opacity(0.32), lineWidth: 2)
-                                    }
-                            }
-                        }
-                    }
-                    .sheet(isPresented: $showMedals) {
-                        StatusModalView(selectedBadge: $selectedbadge)
-                            .presentationDetents( [.height(300), .large], selection: $settingsDetent)
-                    }
-//                    .frame(width: 380, height: 260)
+                    BedgeView()
                     .padding()
                     .overlay {
                         RoundedRectangle(cornerRadius: 20)
