+++ PJ2T12_Again12/PJ2T12_Again12/Views/SwiftUIView.swift
@@ -1,0 +1,18 @@
+//
+//  SwiftUIView.swift
+//  PJ2T12_Again12
+//
+//  Created by kwon ji won on 12/15/23.
+//
+
+import SwiftUI
+
+struct SwiftUIView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    SwiftUIView()
+}
